// src/common/enums/category-message.enum.ts
export enum CategoryMessages {
    NOT_FOUND = 'Category not found',
    ALREADY_EXISTS = 'Category already exists',
    CREATED = 'Category created successfully',
    UPDATED = 'Category updated successfully',
    DELETED = 'Category deleted successfully',

    UNAUTHORIZED_ACCESS = 'Unauthorized access',
    INVALID_CREDENTIALS = 'Invalid credentials',
}